Latest code/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cmudict==1.0.32
colorama==0.4.6
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.1
flask-cors==6.0.1
fuzzywuzzy==0.18.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.4
Levenshtein==0.27.1
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.3.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
preshed==3.0.10
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
pyphen==0.17.2
python-docx==1.1.2
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
